Move delete component
